Added Power implementation to BigDecimal finished Exponent Expansion tests

+++ IrrationalNumbers/IrrationalNumbers.Logic/RemainderResult.cs
@@ -1,11 +1,10 @@
 ﻿
-using IrrationalNumbers.Core;
 
 namespace IrrationalNumbers.Logic
 {
     public class RemainderResult
     {
         public int RemainderOrder { get; set; }
-        public BigFloat Remainder { get; set; }
+        public double Remainder { get; set; }
     }
 }
